// frontend/src/pages/lmEvaluations/types.ts
export type LmModelArgument = {
  name: string;
  url: string;
  tokenizedRequest: boolean;
  tokenizer: string;
};

export type LmEvaluationFormData = {
  deployedModelName: string;
  evaluationName: string;
  tasks: string[];
  modelType: string;
  allowRemoteCode: boolean;
  allowOnline: boolean;
  model: LmModelArgument;
};